+++ googleform-to-backlog/panellers.js
@@ -9,7 +9,7 @@ function send(ev) {
 
   UrlFetchApp.fetch(issuesEndpoint(), {
     method: "post",
-    payload: buildWebsiteIssue(ev.namedValues),
+    payload: buildIssue(ev.namedValues),
   });
 }
 
@@ -24,13 +24,13 @@ function issuesEndpoint() {
 }
 
 /**
- * Build Backlog Issue for website publishing.
+ * Build Backlog Issue for paneller profile registration.
  */
-function buildWebsiteIssue(obj) {
+function buildIssue(obj) {
   const props = PropertiesService.getScriptProperties();
 
   var desc = [
-    "登壇内容を各媒体に掲載お願いします。",
+    "パネルディスカッションの登壇者（パネラー）のプロフィールを掲載お願いします。",
     "\n\n",
     "## 入力内容",
     "\n"
@@ -46,15 +46,9 @@ function buildWebsiteIssue(obj) {
   const todos = [
     "## Website TODOs",
     "",
-    "- [ ] OG画像を作成する --> [Figma](https://www.figma.com/file/gYUjek21b7iucdQtIROyrQ)",
-    "- [ ] Webサイトにセッションを掲載する --> [Shifter](https://go.getshifter.io/admin/teams/181a370f-387e-4615-a1cf-2465145805b6/sites/19430699-e151-4653-af7f-9b58b7d63d95)",
+    "- [ ] Webサイトのパネルセッションを作成/更新する --> [Shifter](https://go.getshifter.io/admin/teams/181a370f-387e-4615-a1cf-2465145805b6/sites/19430699-e151-4653-af7f-9b58b7d63d95)",
     "- [ ] Webサイトのタイムテーブルを更新する --> [Shifter](https://go.getshifter.io/admin/teams/181a370f-387e-4615-a1cf-2465145805b6/sites/19430699-e151-4653-af7f-9b58b7d63d95)",
     "- [ ] Webサイトを公開する",
-    "",
-    "## SNS TODOs",
-    "",
-    "- [ ] BufferでTwitterへの投稿を予約する",
-    "- [ ] BufferでFacebookへの投稿を予約する",
   ].join("\n");
 
   return {
@@ -64,7 +58,7 @@ function buildWebsiteIssue(obj) {
     priorityId: props.getProperty("backlogPriorityId"),
     issueTypeId: props.getProperty("backlogIssueTypeId"),
     assigneeId: props.getProperty("backlogAssigneeId"),
-    summary: "セッションを各媒体に掲載 - " + obj["所属企業・団体"] + " " + obj["登壇者氏名"] + "さん",
+    summary: obj["登壇枠"] + ": パネラーのプロフィールを各媒体に掲載 - " + obj["所属企業・団体"] + " " + obj["登壇者氏名"] + "さん",
     description: [desc.join(""), todos].join("\n"),
   };
 }