+++ googleform-to-backlog/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- Feature: Created panellers to backlog script
 - Feature: Created sessions to backlog script
 - Refactor: Renamed main.js to sponser.js
 - Feature: Created initial script